requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic[email]>=2.8.0
pydantic-settings>=2.3.0
email-validator>=2.0.0
openai>=1.10.0
gspread>=6.0.0
google-auth>=2.26.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0
oauth2client>=4.1.3
APScheduler>=3.10.4
aiohttp>=3.9.1
beautifulsoup4>=4.12.0
python-docx>=1.1.0
docxtpl>=0.16.7
reportlab>=4.0.0
jinja2>=3.1.0
websockets>=12.0

